requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.3
gcloud==0.18.3
googleapis-common-protos==1.51.0
httplib2==0.17.2
idna==2.9
jws==0.1.3
oauth2client==4.1.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.9.7
Pyrebase4==4.3.0
python-jwt==2.0.1
pytz==2019.3
requests==2.23.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
wincertstore==0.2
